Modulated code into different folders
