+++ web/statd.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func main() {
+	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		fmt.Fprint(w, "Hello, World!")
+	})
+	http.ListenAndServe(":8080", nil)
+}
